+++ ChangeLog.md
@@ -4,6 +4,8 @@
 
 ## 1.9
 
+### 1.9.10
+
  * Move eventer SSL debug logging to `debug/eventer/ssl`
  * Fix use-after-free in http logging when HTTP/2 sessions are interrupted.
  * Make the lua subsystem interrupt as it approaches a watchdog timeout.